+++ pytorch-nd-semiconv/src/pytorch_nd_semiconv/_sf_subtract.py
@@ -149,7 +149,7 @@ class SubtractSemifield(NamedTuple):
 
         Parameters
         ----------
-        p : int
+        p : float
             The power to use in \(\oplus_p\).
             May not be zero.
         """
@@ -182,7 +182,7 @@ class SubtractSemifield(NamedTuple):
 
         Parameters
         ----------
-        mu : int
+        mu : float
             The base to use in \(\oplus_\mu\).
             May not be zero.
         """
